Scoping (nearly) all styles Moving the version widget to the footer Reworking the footer flexbox
