+++ src/modules/drawing/DrawingModule.vue
@@ -283,7 +283,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 .tooltip-measure,
 .draw-measure-tmp,
 .draw-help-popup {
