+++ src/modules/drawing/components/ProfilePopup.vue
@@ -345,7 +345,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/bootstrap-theme';
 
 .profile-popup {
