+++ src/modules/menu/MenuModule.vue
@@ -9,7 +9,6 @@
                 <slot />
                 <HeaderMenuButton />
                 <MenuTray class="menu-tray" />
-                <div v-if="showAppVersion" class="app-version small">v{{ appVersion }}</div>
             </div>
         </div>
     </transition>
@@ -24,7 +23,6 @@ import HeaderSwissConfederationText from './components/header/HeaderSwissConfede
 
 import MenuTray from './components/MenuTray'
 import HeaderLoadingBar from '@/modules/menu/components/header/HeaderLoadingBar'
-import { APP_VERSION, DEBUG } from '@/config'
 
 export default {
     components: {
@@ -34,24 +32,16 @@ export default {
         SwissFlag,
         MenuTray,
     },
-    data() {
-        return {
-            appVersion: APP_VERSION,
-        }
-    },
     computed: {
         ...mapState({
             showHeader: (state) => state.ui.showHeader,
             showLoadingBar: (state) => state.ui.showLoadingBar,
         }),
-        showAppVersion: function () {
-            return DEBUG
-        },
     },
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/media-query.mixin';
 
 .header {
@@ -81,13 +71,6 @@ export default {
             width: 95%;
             max-width: 40rem;
         }
-    }
-    .app-version {
-        position: fixed;
-        bottom: 0.1rem;
-        left: 50%;
-        transform: translate(-50%, 0);
-        color: $gray-400;
     }
 }
 .slide-up-leave-active,
