+++ src/modules/menu/components/MenuSection.vue
@@ -60,7 +60,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 .menu-section-head {
     display: flex;
     cursor: pointer;
