+++ src/modules/menu/components/activeLayers/MenuActiveLayersListItem.vue
@@ -171,7 +171,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/bootstrap-theme';
 @import '../../scss/menu-items';
 .menu-layer-list-item {
