+++ src/modules/menu/components/header/SwissFlag.vue
@@ -13,7 +13,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/bootstrap-theme';
 .swiss-flag {
     height: 2em;
