+++ src/modules/menu/components/topics/MenuTopicSelectionPopup.vue
@@ -37,7 +37,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/variables';
 .menu-topic-selector-popup {
     * {
