+++ src/modules/menu/components/topics/MenuTopicTreeItem.vue
@@ -90,7 +90,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/bootstrap-theme';
 @import '../../scss/menu-items';
 .menu-topic-tree-item {
