+++ src/modules/menu/components/topics/TopicIcon.vue
@@ -17,6 +17,6 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import '../../scss/topics-image';
 </style>
