+++ src/modules/overlay/OverlayModule.vue
@@ -26,7 +26,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/variables';
 #overlay {
     position: fixed;
