+++ src/modules/overlay/components/ModalWithOverlay.vue
@@ -93,7 +93,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/variables';
 .modal-popup {
     position: fixed;
