+++ src/modules/search/SearchModule.vue
@@ -18,7 +18,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 #search {
     width: 100%;
 }
