+++ src/modules/search/components/SearchResultCategory.vue
@@ -43,7 +43,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/bootstrap-theme';
 .search-category {
     .search-category-header {
