+++ src/modules/search/components/SearchResultListEntry.vue
@@ -34,7 +34,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/bootstrap-theme';
 .search-category-entry {
     @extend .p-2;
