+++ src/modules/toolbox/ToolboxModule.vue
@@ -27,7 +27,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/media-query.mixin';
 
 #right-toolbox,
