+++ src/modules/toolbox/components/BackgroundSelectorButton.vue
@@ -73,7 +73,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/bootstrap-theme';
 
 $bg-button-border-size: 8px;
