+++ src/modules/toolbox/components/GeolocButton.vue
@@ -26,7 +26,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/bootstrap-theme';
 
 $normal-color: $gray-800;
