+++ src/modules/toolbox/components/ZoomButtons.vue
@@ -29,7 +29,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/bootstrap-theme';
 $normal-color: $gray-800;
 $active-color: $primary;
