+++ src/modules/tooltip/components/TooltipBox.vue
@@ -44,7 +44,7 @@ export default {
 }
 </script>
 
-<style lang="scss">
+<style lang="scss" scoped>
 @import 'src/scss/variables';
 .tooltip-box {
     position: absolute;
